+++ tardis/sftp/apps.py
@@ -1,0 +1,6 @@
+from tardis.app_config import AbstractTardisAppConfig
+
+
+class SFTPConfig(AbstractTardisAppConfig):
+    name = 'tardis.sftp'
+    verbose_name = 'SFTP'
