+++ tardis/sftp/management/commands/sftpd.py
@@ -3,7 +3,7 @@ import logging
 from django.conf import settings
 from django.core.management.base import BaseCommand
 
-import tardis.tardis_portal.sftp as sftp
+from tardis.sftp import sftp
 
 logger = logging.getLogger(__name__)
