+++ tardis/tardis_portal/templates/tardis_portal/ajax/experiment_description.html
@@ -126,7 +126,7 @@
         </a>
       {% endfor %}
     {% endif %}
-    <a href="{% url "tardis.tardis_portal.views.sftp_access" %}?object_id={{ experiment.id }}&object_type=experiment"
+    <a href="{% url "tardis.sftp:index" %}?object_id={{ experiment.id }}&object_type=experiment"
        target="_blank" class="btn btn-primary btn-mini" title="Download with SFTP">SFTP</a>
     </dd>
   </dl>
