+++ tardis/tardis_portal/views/images.py
@@ -6,7 +6,6 @@ from base64 import b64decode
 import logging
 from os import path
 
-from PIL import Image, ImageFont, ImageDraw
 from django.conf import settings
 from django.contrib.auth.decorators import login_required
 from django.http import HttpResponse, HttpResponseNotFound, HttpResponseRedirect
@@ -18,56 +17,6 @@ from ..models import ExperimentParameter, DatasetParameter, \
 from ..shortcuts import return_response_error
 
 logger = logging.getLogger(__name__)
-
-
-@login_required
-def cybderduck_connection_window(request):
-    base_image = ("tardis/tardis_portal/templates/images/"
-                  "cyberduck_connection_blank.png")
-    font_file = "tardis/tardis_portal/templates/fonts/roboto.ttf"
-    base = Image.open(base_image)
-    font = ImageFont.truetype(font_file, 13)
-    draw = ImageDraw.Draw(base)
-
-    if request.user.userprofile.isDjangoAccount:
-        sftp_username = request.user.username
-    else:
-        login_attr = getattr(settings, 'SFTP_USERNAME_ATTRIBUTE', 'email')
-        sftp_username = getattr(request.user, login_attr)
-
-    sftp_host = request.get_host().split(':')[0]
-    sftp_port = str(getattr(settings, 'SFTP_PORT', 2200))
-    info = [
-        {'location': (247, 170),
-         'text': sftp_host},
-        {'location': (530, 169),
-         'text': sftp_port},
-        {'location': (247, 217),
-         'text': sftp_username},
-    ]
-    url = {'location': (241, 195),
-           'text': 'sftp://{}@{}:{}/'.format(
-               sftp_username,
-               sftp_host,
-               sftp_port)}
-    for text in info:
-        draw.text(text['location'], text['text'], font=font,
-                  fill=(0, 0, 0))
-
-    def draw_underlined_text(draw, pos, text, font, **options):
-        twidth, theight = draw.textsize(text, font=font)
-        lx, ly = pos[0], pos[1] + theight - 2
-        draw.text(pos, text, font=font, **options)
-        draw.line((lx, ly, lx + twidth - 2, ly), **options)
-
-    url_font = ImageFont.truetype(font_file, 11)
-    url_colour = (0, 49, 249)
-    draw_underlined_text(draw, url['location'],
-                         url['text'], url_font, fill=url_colour)
-    response = HttpResponse(content_type='image/png')
-    base.save(response, "PNG")
-    base.save("foo.png")
-    return response
 
 
 @never_cache
