+++ tardis/tardis_portal/views/pages.py
@@ -596,70 +596,6 @@ def user_guide(request):
 
 
 @login_required
-def sftp_access(request):
-    """
-    Show dynamically generated instructions on how to connect to SFTP
-    :param Request request: HttpRequest
-    :return: HttpResponse
-    :rtype: HttpResponse
-    """
-    from ..download import make_mapper
-    object_type = request.GET.get('object_type')
-    object_id = request.GET.get('object_id')
-    sftp_start_dir = ''
-    if object_type and object_id:
-        ct = ContentType.objects.get_by_natural_key(
-            'tardis_portal', object_type)
-        item = ct.model_class().objects.get(id=object_id)
-        if object_type == 'experiment':
-            exps = [item]
-            dataset = None
-            datafile = None
-        else:
-            if object_type == 'dataset':
-                dataset = item
-                datafile = None
-            elif object_type == 'datafile':
-                datafile = item
-                dataset = datafile.dataset
-            exps = dataset.experiments.all()
-        allowed_exps = []
-        for exp in exps:
-            if has_experiment_download_access(request, exp.id):
-                allowed_exps.append(exp)
-        if allowed_exps:
-            path_mapper = make_mapper(settings.DEFAULT_PATH_MAPPER,
-                                      rootdir=None)
-            exp = allowed_exps[0]
-            path_parts = ['/home', request.user.username, 'experiments',
-                          path_mapper(exp)]
-            if dataset is not None:
-                path_parts.append(path_mapper(dataset))
-            if datafile is not None:
-                path_parts.append(datafile.directory)
-            sftp_start_dir = path.join(*path_parts)
-
-    if request.user.userprofile.isDjangoAccount:
-        sftp_username = request.user.username
-    else:
-        login_attr = getattr(settings, 'SFTP_USERNAME_ATTRIBUTE', 'email')
-        sftp_username = getattr(request.user, login_attr)
-    c = {
-        'sftp_host': request.get_host().split(':')[0],
-        'sftp_port': getattr(settings, 'SFTP_PORT', 2200),
-        'sftp_username': sftp_username,
-        'sftp_start_dir': sftp_start_dir,
-        'site_name': getattr(settings, 'SITE_TITLE', 'MyTardis'),
-    }
-    c['sftp_url'] = 'sftp://{}@{}:{}{}'.format(
-        c['sftp_username'],
-        c['sftp_host'],
-        c['sftp_port'],
-        c['sftp_start_dir'])
-    return render(request, template_name='tardis_portal/sftp.html', context=c)
-
-
-@login_required
 def facility_overview(request):
     '''
     summary of experiments in a facility
