+++ tardis/urls/__init__.py
@@ -94,6 +94,9 @@ urlpatterns = [
     # Facility views
     url(r'^facility/', include(facility_urls)),
 
+    # SFTP views
+    (r'^sftp/', include('tardis.sftp.urls', namespace='sftp', app_name='tardis.sftp')),
+
     # Display Views
     url(r'^display/', include(display_urls)),
